ajax call works...

+++ src/main/resources/db/migration/V1.3__orderfix.sql
@@ -1,2 +1,2 @@
 alter table orders
-add column item_price decimal(50);
+add column item_price double(6,2);
